feat: add manual update check to manager page
This commit introduces a manual update check feature in the manager page. Users can now tap a "Check Update" button to trigger an immediate check for updates. The update check will also now notify the user with a snackbar if no updates are available when it's manually triggered. This change modifies `update_available.dart` to allow force triggering the update and adds a new ListTile to `manager_page.dart` to house the button.